Move console accessors to recipes.utils.rich
