+++ src/fairseq2/recipes/llama/convert_checkpoint.py
@@ -22,7 +22,7 @@ from fairseq2.logging import get_log_writer
 from fairseq2.models.llama import load_llama_config
 from fairseq2.models.llama.integ import convert_to_reference_checkpoint
 from fairseq2.recipes.cli import CliCommandHandler
-from fairseq2.recipes.console import get_error_console
+from fairseq2.recipes.utils.rich import get_error_console
 from fairseq2.utils.file import dump_torch_tensors, load_torch_tensors
 
 log = get_log_writer(__name__)
