+++ src/fairseq2/recipes/utils/rich.py
@@ -6,6 +6,8 @@
 
 from __future__ import annotations
 
+from rich import get_console as get_rich_console
+from rich.console import Console
 from rich.progress import (
     BarColumn,
     Progress,
@@ -19,7 +21,41 @@ from rich.text import Text
 from typing_extensions import override
 
 from fairseq2.gang import get_rank
-from fairseq2.recipes.console import get_error_console
+
+_console: Console | None = None
+
+
+def get_console() -> Console:
+    global _console
+
+    if _console is None:
+        _console = get_rich_console()
+
+    return _console
+
+
+def set_console(console: Console) -> None:
+    global _console
+
+    _console = console
+
+
+_error_console: Console | None = None
+
+
+def get_error_console() -> Console:
+    global _error_console
+
+    if _error_console is None:
+        _error_console = Console(stderr=True, highlight=False)
+
+    return _error_console
+
+
+def set_error_console(console: Console) -> None:
+    global _error_console
+
+    _error_console = console
 
 
 def create_rich_progress() -> Progress:
